lastvwtheme
verthem

+++ update/dev/tests/integration/testsuite/Magento/Update/RollbackTest.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * Copyright © 2015 Magento. All rights reserved.
+ * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
 namespace Magento\Update;
 
-class RollbackTest extends \PHPUnit_Framework_TestCase
+class RollbackTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @var \Magento\Update\Rollback
@@ -61,7 +61,9 @@ class RollbackTest extends \PHPUnit_Framework_TestCase
 
         $this->backupFileName = $this->backupPath . '/../' . uniqid() . '_code.tar';
         $this->backupFile = new \PharData($this->backupFileName);
-        $this->backupInfo = $this->getMock('Magento\Update\Backup\BackupInfo', [], [], '', false);
+        $this->backupInfo = $this->getMockBuilder(\Magento\Update\Backup\BackupInfo::class)
+            ->disableOriginalConstructor()
+            ->getMock();
         $this->rollBack = new \Magento\Update\Rollback($this->backupPath, $this->archivedDir, null, $this->backupInfo);
     }
 
@@ -77,9 +79,18 @@ class RollbackTest extends \PHPUnit_Framework_TestCase
             unlink($gtzFile);
         }
 
-        rmdir($this->backupPath);
-        rmdir($this->excludedDir);
-        rmdir($this->archivedDir);
+        if (is_dir($this->backupPath)) {
+            rmdir($this->backupPath);
+        }
+        if (is_dir($this->excludedDir)) {
+            rmdir($this->excludedDir);
+        }
+        if (is_dir($this->archivedDir)) {
+            rmdir($this->archivedDir);
+        }
+        if (file_exists(MAGENTO_BP . '/var/.update_status.txt')) {
+            unlink(MAGENTO_BP . '/var/.update_status.txt');
+        }
     }
 
     /**
@@ -102,7 +113,7 @@ class RollbackTest extends \PHPUnit_Framework_TestCase
         $newFile = $this->backupPath . '/' . uniqid() . '_code.tgz';
         copy($this->backupFileName, $newFile);
         if (file_exists($this->backupFileName)) {
-            unset($this->backupFile);
+            $this->backupFile = null;
             unlink($this->backupFileName);
         }
         $gtzFile = str_replace('tar', 'tgz', $this->backupFileName);
